fix LGTM warnings (#479)
* fix LGTM warnings
* fix vs warnings
* fix some codacy warnings

+++ modules/f2c/src/c/libf2c/lwrite.c
@@ -202,7 +202,7 @@ l_g(char *buf, double n)
                 }
     }
 f__ret:
-    return b - buf;
+    return (int)(b - buf);
 #endif
 }
